Windows 64-bit support (#316)...
Minor fixes to tools.

Add cef.GetVersion() function with keys like cef_version, chrome_version etc.

Print architecture along with python version in examples/tools.

+++ src/window_utils_win.pyx
@@ -117,7 +117,8 @@ class WindowUtils(object):
             # of the parent window, try again.
 
             if not parentIconBig and not parentIconSmall:
-                parentWindowHandle = <int>GetParent(<HWND>parentWindowHandle)
+                parentWindowHandle = <uintptr_t>GetParent(
+                                                    <HWND>parentWindowHandle)
 
             Debug("WindowUtils.SetIcon(): popup inherits icon from "
                     "parent window: %s" % parentWindowHandle)
